requirements.txt
python-dotenv
langchain-google-genai
langchain-community
faiss-cpu
langchain